Реализовал считывание контактов из тел.книги
+ фикс багов имя пользователя удаляло всю ноду имен и базы данных (переписал проверку на удаление)
+ добавил статус log out, при выходе пользователя из приложения

+++ app/src/main/java/com/evgeny_m/messenger3/utils/firebase.kt
@@ -2,6 +2,7 @@ package com.evgeny_m.messenger3.utils
 
 import android.annotation.SuppressLint
 import android.net.Uri
+import android.provider.ContactsContract
 import android.util.Log
 import android.widget.ImageView
 import android.widget.TextView
@@ -13,6 +14,7 @@ import com.evgeny_m.messenger3.fragments.main.settings.backToSettingsFragment
 import com.evgeny_m.messenger3.fragments.main.settings.settings_fragments.ChangeUserNameFragment
 import com.evgeny_m.messenger3.fragments.main.settings.showToast
 import com.evgeny_m.messenger3.fragments.register.replacePhoneNumber
+import com.evgeny_m.messenger3.model.CommonModel
 import com.evgeny_m.messenger3.model.UserModel
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.database.DatabaseReference
@@ -26,20 +28,26 @@ lateinit var storage: StorageReference
 lateinit var currentUserId: String
 lateinit var user: UserModel
 
+private const val LOGING = false
+private const val LOG_TAG = "firebase"
+
 /**
  *  Статические поля для NavDrawer, что бы обновлять данные с помощъю ChatsFragment
  *  они проходят инициализацию в NavHeaderFragment в методе onStart
  */
 
 @SuppressLint("StaticFieldLeak")
-lateinit var userPhone : TextView
+lateinit var userPhone: TextView
+
 @SuppressLint("StaticFieldLeak")
-lateinit var userFullName : TextView
+lateinit var userFullName: TextView
+
 @SuppressLint("StaticFieldLeak")
-lateinit var userPhoto : ImageView
+lateinit var userPhoto: ImageView
 
 const val NODE_USERS = "users"
 const val NODE_USER_NAMES = "usernames"
+const val NODE_PHONES = "users_phones"
 
 const val FOLDER_PROFILE_IMAGE = "profile_image"
 
@@ -51,9 +59,7 @@ const val CHILD_BIO = "bio"
 const val CHILD_USER_PHOTO = "photoUrl"
 const val CHILD_STATUS = "status"
 
-
-
-fun initFirebase(){
+fun initFirebase() {
     auth = FirebaseAuth.getInstance()
     database = FirebaseDatabase.getInstance().reference
     storage = FirebaseStorage.getInstance().reference
@@ -71,12 +77,12 @@ fun initUser() {
     database
         .child(NODE_USERS)
         .child(currentUserId)
-        .addValueEventListener(AppValueEventListener{
+        .addValueEventListener(AppValueEventListener {
             user = it.getValue(UserModel::class.java) ?: UserModel()
         })
 }
 
-fun saveUserData(inputUserData: String, child : String ) {
+fun saveUserData(inputUserData: String, child: String) {
     database
         .child(NODE_USERS)
         .child(currentUserId)
@@ -106,12 +112,16 @@ fun SettingsFragment.readUserData() {
         .child(currentUserId)
         .child(CHILD_PHONE)
         .get().addOnSuccessListener {
-            Log.d("firebase", "Got value ${it.value}")
+            if (LOGING) {
+                Log.d(LOG_TAG, "Got value ${it.value}")
+            }
             userPhone = binding.settingsUserPhone
             userPhone.text = it.value as CharSequence?
 
         }.addOnFailureListener {
-            Log.d("firebase", "Error getting data", it)
+            if (LOGING) {
+                Log.d(LOG_TAG, "Error getting data", it)
+            }
         }
     /**
      * Читаем имя пользователя и передаем в Поле имени во фрагменте настроек
@@ -121,12 +131,16 @@ fun SettingsFragment.readUserData() {
         .child(currentUserId)
         .child(CHILD_USERFULLNAME)
         .get().addOnSuccessListener {
-            Log.d("firebase", "Got value ${it.value}")
+            if (LOGING) {
+                Log.d(LOG_TAG, "Got value ${it.value}")
+            }
             userFullName = binding.settingsUserFullName
             userFullName.text = it.value as CharSequence?
 
         }.addOnFailureListener {
-            Log.d("firebase", "Error getting data", it)
+            if (LOGING) {
+                Log.d(LOG_TAG, "Error getting data", it)
+            }
         }
 
     /**
@@ -137,12 +151,17 @@ fun SettingsFragment.readUserData() {
         .child(currentUserId)
         .child(CHILD_USERNAME)
         .get().addOnSuccessListener {
-            Log.d("firebase", "Got value ${it.value}")
+            if (LOGING) {
+                Log.d(LOG_TAG, "Got value ${it.value}")
+            }
             userName = binding.settingsUserName
             userName.text = "@${it.value as CharSequence?}"
 
         }.addOnFailureListener {
-            Log.d("firebase", "Error getting data", it)
+            if (LOGING) {
+                Log.d(LOG_TAG, "Error getting data", it)
+            }
+
         }
 
     /**
@@ -153,24 +172,30 @@ fun SettingsFragment.readUserData() {
         .child(currentUserId)
         .child(CHILD_BIO)
         .get().addOnSuccessListener {
-            Log.d("firebase", "Got value ${it.value}")
+            if (LOGING) {
+                Log.d(LOG_TAG, "Got value ${it.value}")
+            }
             userBio = binding.settingsUserBio
             userBio.text = it.value as CharSequence?
 
         }.addOnFailureListener {
-            Log.d("firebase", "Error getting data", it)
+            if (LOGING) {
+                Log.d(LOG_TAG, "Error getting data", it)
+            }
         }
 
     /**
      * Читаем photoURL пользователя и передаем обновляем фото пользователя во фрагменте настроек
      */
 
-  database
+    database
         .child(NODE_USERS)
         .child(currentUserId)
         .child(CHILD_USER_PHOTO)
         .get().addOnSuccessListener {
-            Log.d("firebase", "Got value ${it.value}")
+            if (LOGING) {
+                Log.d(LOG_TAG, "Got value ${it.value}")
+            }
             userPhoto = binding.settingsUserImage
             Glide
                 .with(this)
@@ -178,8 +203,10 @@ fun SettingsFragment.readUserData() {
                 .centerCrop()
                 .into(userPhoto)
 
-        }.addOnFailureListener{
-            Log.d("firebase", "Error getting data", it)
+        }.addOnFailureListener {
+            if (LOGING) {
+                Log.d(LOG_TAG, "Error getting data", it)
+            }
         }
 }
 
@@ -194,7 +221,7 @@ fun SettingsFragment.readUserData() {
 fun ChangeUserNameFragment.checkingAndAddUserName() {
 
     database
-        .child(NODE_USER_NAMES).addListenerForSingleValueEvent(AppValueEventListener{
+        .child(NODE_USER_NAMES).addListenerForSingleValueEvent(AppValueEventListener {
             if (it.hasChild(newUserName)) {
                 showToast("name is taken")
             } else {
@@ -203,9 +230,22 @@ fun ChangeUserNameFragment.checkingAndAddUserName() {
         })
 }
 
+/**
+ * checkOldUserName() - проверяет принадлежность имени к данному ID пользователя.
+ * если имени нет создает новое имя в базе, если имя есть и его ID совпадает с ID пользователя
+ * то добовляет новое имя и удаляет старое из базы данных
+ */
+
 private fun checkOldUserName(oldUserName: String, newUserName: String) {
-    database.child(NODE_USER_NAMES).addListenerForSingleValueEvent(AppValueEventListener{
-        if (it.hasChild(oldUserName)) {
+    database.child(NODE_USER_NAMES).addListenerForSingleValueEvent(AppValueEventListener {
+        if (LOGING) {
+            Log.d(LOG_TAG, "Got value ${it.value}")
+            Log.d(LOG_TAG, "Got value ${it.hasChild(oldUserName)}")
+            Log.d(LOG_TAG, "Got value ${it.child(oldUserName).value}")
+            Log.d(LOG_TAG, "Got value $currentUserId")
+        }
+
+        if (it.hasChild(oldUserName) && it.child(oldUserName).value == currentUserId) {
             deleteOldUserName(oldUserName, newUserName)
         } else {
             addNewUserName(newUserName)
@@ -251,17 +291,18 @@ private fun updateCurrentUserName(newUserName: String) {
  */
 
 fun readUserDataToNavHeader() {
-
     database
         .child(NODE_USERS)
         .child(currentUserId)
         .child(CHILD_PHONE)
         .get().addOnSuccessListener {
-            Log.d("firebase", "Got value ${it.value}")
+            if (LOGING) {
+                Log.d(LOG_TAG, "Got value ${it.value}")
+            }
             userPhone.text = it.value as CharSequence?
 
-        }.addOnFailureListener{
-            Log.d("firebase", "Error getting data", it)
+        }.addOnFailureListener {
+            Log.d(LOG_TAG, "Error getting data", it)
         }
 
     database
@@ -269,11 +310,13 @@ fun readUserDataToNavHeader() {
         .child(currentUserId)
         .child(CHILD_USERFULLNAME)
         .get().addOnSuccessListener {
-            Log.d("firebase", "Got value ${it.value}")
+            if (LOGING) {
+                Log.d(LOG_TAG, "Got value ${it.value}")
+            }
             userFullName.text = it.value as CharSequence?
 
-        }.addOnFailureListener{
-            Log.d("firebase", "Error getting data", it)
+        }.addOnFailureListener {
+            Log.d(LOG_TAG, "Error getting data", it)
         }
 
     database
@@ -281,7 +324,9 @@ fun readUserDataToNavHeader() {
         .child(currentUserId)
         .child(CHILD_USER_PHOTO)
         .get().addOnSuccessListener {
-            Log.d("firebase", "Got value ${it.value}")
+            if (LOGING) {
+                Log.d(LOG_TAG, "Got value ${it.value}")
+            }
 
             Glide
                 .with(APP_ACTIVITY)
@@ -289,9 +334,8 @@ fun readUserDataToNavHeader() {
                 .centerCrop()
                 .into(userPhoto)
 
-
-        }.addOnFailureListener{
-            Log.d("firebase", "Error getting data", it)
+        }.addOnFailureListener {
+            Log.d(LOG_TAG, "Error getting data", it)
         }
 }
 
@@ -304,21 +348,34 @@ fun Fragment.authNewUser() {
     val uid = auth.currentUser?.uid.toString()
     dateMap[CHILD_ID] = uid
     dateMap[CHILD_PHONE] = replacePhoneNumber
-    dateMap[CHILD_USERFULLNAME] = uid
-    dateMap[CHILD_USERNAME] = uid
 
     database
-        .child(NODE_USERS)
-        .child(uid)
-        .updateChildren(dateMap)
-        .addOnCompleteListener { task2 ->
-            if (task2.isSuccessful) {
-                showToast("авторизация прошла")
-                replaceActivity(MainActivity())
-            } else {
-                showToast("авторизация прошла но данные не записаны")
-                Log.d ("TASK2" , task2.exception?.message.toString())
+        .child(NODE_PHONES)
+        .child(replacePhoneNumber)
+        .setValue(uid)
+        .addOnSuccessListener {
+            database
+                .child(NODE_USERS)
+                .child(uid)
+                .updateChildren(dateMap)
+                .addOnSuccessListener {
+
+                    showToast("авторизация прошла")
+                    replaceActivity(MainActivity())
+
+                }.addOnFailureListener {
+
+                    if (LOGING) {
+                        Log.d(LOG_TAG, it.message.toString())
+                    }
+                    showToast("авторизация не прошла")
+                }
+        }.addOnFailureListener {
+
+            if (LOGING) {
+                Log.d(LOG_TAG, it.message.toString())
             }
+            showToast("авторизация не прошла")
         }
 }
 
@@ -347,6 +404,41 @@ fun saveUserPhotoUrl(photoUrl: String, childUserPhoto: String) {
                 showToast("data not saved")
             }
         }
+}
+
+/**
+ * initContacts() - метод считывает контакты в массив и одрабатывает номера телефонов
+ * для удобной работы с базой данных
+ */
+
+fun initContacts() {
+    if (checkPermission(READ_CONTACTS)) {
+        val arrayContacts = arrayListOf<CommonModel>()
+        val cursor = APP_ACTIVITY.contentResolver.query(
+            ContactsContract.CommonDataKinds.Phone.CONTENT_URI,
+            null,
+            null,
+            null,
+            null
+        )
+        cursor?.let {
+            while (cursor.moveToNext()) {
+                val fullname =
+                    it.getString(it.getColumnIndex(ContactsContract.Contacts.DISPLAY_NAME))
+                val phone =
+                    it.getString(it.getColumnIndex(ContactsContract.CommonDataKinds.Phone.NUMBER))
+                val newModel = CommonModel()
+
+                newModel.userfullname = fullname
+                newModel.phone = phone.replace(Regex("[\\s,()-]"), "")
+                if (newModel.phone[0].toString() == "8") {
+                    newModel.phone = newModel.phone.replaceFirst("8", "+7")
+                }
+                arrayContacts.add(newModel)
+            }
+        }
+        cursor?.close()
+    }
 }
